+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/PhysicalReaderUtil.java
@@ -21,6 +21,7 @@ package io.pixelsdb.pixels.common.physical;
 
 import io.pixelsdb.pixels.common.physical.io.PhysicalHDFSReader;
 import io.pixelsdb.pixels.common.physical.io.PhysicalLocalReader;
+import io.pixelsdb.pixels.common.physical.io.PhysicalMinIOReader;
 import io.pixelsdb.pixels.common.physical.io.PhysicalS3Reader;
 
 import java.io.IOException;
@@ -53,6 +54,9 @@ public class PhysicalReaderUtil
                 break;
             case s3:
                 reader = new PhysicalS3Reader(storage, path);
+                break;
+            case minio:
+                reader = new PhysicalMinIOReader(storage, path);
                 break;
             default:
                 throw new IOException("Storage scheme '" + storage.getScheme() + "' is not supported.");
