+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/io/AbstractS3Reader.java
@@ -1,0 +1,253 @@
+/*
+ * Copyright 2022 PixelsDB.
+ *
+ * This file is part of Pixels.
+ *
+ * Pixels is free software: you can redistribute it and/or modify
+ * it under the terms of the Affero GNU General Public License as
+ * published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * Pixels is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * Affero GNU General Public License for more details.
+ *
+ * You should have received a copy of the Affero GNU General Public
+ * License along with Pixels.  If not, see
+ * <https://www.gnu.org/licenses/>.
+ */
+package io.pixelsdb.pixels.common.physical.io;
+
+import io.pixelsdb.pixels.common.physical.PhysicalReader;
+import io.pixelsdb.pixels.common.physical.Storage;
+import io.pixelsdb.pixels.common.physical.storage.AbstractS3;
+import io.pixelsdb.pixels.common.utils.ConfigFactory;
+import software.amazon.awssdk.core.ResponseBytes;
+import software.amazon.awssdk.core.sync.ResponseTransformer;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.GetObjectRequest;
+import software.amazon.awssdk.services.s3.model.GetObjectResponse;
+
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicLong;
+
+/**
+ * The abstract class for the physical readers of AWS S3 compatible storage systems.
+ *
+ * @author hank
+ * Created at: 10/04/2022
+ */
+public abstract class AbstractS3Reader implements PhysicalReader
+{
+    /*
+     * The implementations of most methods in this class are from its subclass PhysicalS3Reader.
+     */
+
+    protected static boolean enableAsync = false;
+    protected static boolean useAsyncClient = false;
+    protected static final ExecutorService clientService;
+
+    static
+    {
+        int clientServiceThreads = Integer.parseInt(
+                ConfigFactory.Instance().getProperty("s3.client.service.threads"));
+        ThreadGroup clientServiceGroup = new ThreadGroup("s3.client.service");
+        clientServiceGroup.setDaemon(true);
+        clientServiceGroup.setMaxPriority(Thread.MAX_PRIORITY-1);
+        clientService = Executors.newFixedThreadPool(clientServiceThreads, runnable -> {
+            Thread thread = new Thread(clientServiceGroup, runnable);
+            thread.setDaemon(true);
+            /**
+             * Issue #136:
+             * Using MAX_PRIORITY-1 for client threads accelerates async read with sync client for
+             * some queries that send many get object requests (e.g. 5%).
+             * Note that clientService is only used for the cause that async client is disabled while
+             * async read is enabled.
+             */
+            thread.setPriority(Thread.MAX_PRIORITY-1);
+            return thread;
+        });
+
+        Runtime.getRuntime().addShutdownHook(new Thread(clientService::shutdownNow));
+    }
+
+    protected final AbstractS3 s3;
+    protected final AbstractS3.Path path;
+    protected final String pathStr;
+    protected final long id;
+    protected final AtomicLong position;
+    protected final long length;
+    protected final S3Client client;
+
+    public AbstractS3Reader(Storage storage, String path) throws IOException
+    {
+        // instanceof tells if an object is an instance of a class or its parent classes.
+        if (storage instanceof AbstractS3)
+        {
+            this.s3 = (AbstractS3) storage;
+        }
+        else
+        {
+            throw new IOException("Storage is not S3 compatible.");
+        }
+        if (path.contains("://"))
+        {
+            // remove the scheme.
+            path = path.substring(path.indexOf("://") + 3);
+        }
+        this.path = new AbstractS3.Path(path);
+        this.pathStr = path;
+        this.id = this.s3.getFileId(path);
+        this.length = this.s3.getStatus(path).getLength();
+        this.position = new AtomicLong(0);
+        this.client = this.s3.getClient();
+
+        enableAsync = Boolean.parseBoolean(ConfigFactory.Instance().getProperty("s3.enable.async"));
+        useAsyncClient = Boolean.parseBoolean(ConfigFactory.Instance().getProperty("s3.use.async.client"));
+    }
+
+    protected String toRange(long start, int length)
+    {
+        assert start >= 0 && length > 0;
+        StringBuilder builder = new StringBuilder("bytes=");
+        builder.append(start).append('-').append(start+length-1);
+        return builder.toString();
+    }
+
+    @Override
+    public long getFileLength() throws IOException
+    {
+        return length;
+    }
+
+    @Override
+    public void seek(long desired) throws IOException
+    {
+        if (0 <= desired && desired < length)
+        {
+            position.set(desired);
+            return;
+        }
+        throw new IOException("Desired offset " + desired +
+                " is out of bound (" + 0 + "," + length + ")");
+    }
+
+    @Override
+    public ByteBuffer readFully(int len) throws IOException
+    {
+        if (this.position.get() + len > this.length)
+        {
+            throw new IOException("Current position " + this.position.get() + " plus " +
+                    len + " exceeds object length " + this.length + ".");
+        }
+        GetObjectRequest request = GetObjectRequest.builder().bucket(path.bucket)
+                .key(path.key).range(toRange(position.get(), len)).build();
+        ResponseBytes<GetObjectResponse> response =
+                client.getObject(request, ResponseTransformer.toBytes());
+        try
+        {
+            this.position.addAndGet(len);
+            return ByteBuffer.wrap(response.asByteArrayUnsafe());
+        } catch (Exception e)
+        {
+            throw new IOException("Failed to read object.", e);
+        }
+    }
+
+    @Override
+    public void readFully(byte[] buffer) throws IOException
+    {
+        ByteBuffer byteBuffer = readFully(buffer.length);
+        System.arraycopy(byteBuffer.array(), 0, buffer, 0, buffer.length);
+    }
+
+    @Override
+    public void readFully(byte[] buffer, int off, int len) throws IOException
+    {
+        ByteBuffer byteBuffer = readFully(len);
+        System.arraycopy(byteBuffer.array(), 0, buffer, off, len);
+    }
+
+    /**
+     * @return true if readAsync is supported.
+     */
+    @Override
+    public boolean supportsAsync()
+    {
+        return enableAsync;
+    }
+
+    @Override
+    abstract public CompletableFuture<ByteBuffer> readAsync(long offset, int len) throws IOException;
+
+    @Override
+    public long readLong() throws IOException
+    {
+        ByteBuffer buffer = readFully(Long.BYTES);
+        return buffer.getLong();
+    }
+
+    @Override
+    public int readInt() throws IOException
+    {
+        ByteBuffer buffer = readFully(Integer.BYTES);
+        return buffer.getInt();
+    }
+
+    @Override
+    abstract public void close() throws IOException;
+
+    @Override
+    public String getPath()
+    {
+        return pathStr;
+    }
+
+    /**
+     * Get the last domain in path.
+     *
+     * @return
+     */
+    @Override
+    public String getName()
+    {
+        return path.key;
+    }
+
+    /**
+     * For a file or object in the storage, it may have one or more
+     * blocks. Each block has its unique id. This method returns the
+     * block id of the current block that is been reading.
+     * <p>
+     * For local fs, each file has only one block id, which is also
+     * the file id.
+     *
+     * <p>Note: Storage.getFileId() assumes that each file or object
+     * only has one block. In this case, the file id is the same as
+     * the block id here.</p>
+     *
+     * @return
+     * @throws IOException
+     */
+    @Override
+    public long getBlockId() throws IOException
+    {
+        return this.id;
+    }
+
+    /**
+     * Get the scheme of the backed physical storage.
+     *
+     * @return
+     */
+    @Override
+    public Storage.Scheme getStorageScheme()
+    {
+        return s3.getScheme();
+    }
+}
