+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/storage/S3.java
@@ -19,72 +19,50 @@
  */
 package io.pixelsdb.pixels.common.physical.storage;
 
-import io.etcd.jetcd.KeyValue;
 import io.pixelsdb.pixels.common.exception.StorageException;
-import io.pixelsdb.pixels.common.physical.Status;
-import io.pixelsdb.pixels.common.physical.Storage;
-import io.pixelsdb.pixels.common.physical.io.S3InputStream;
-import io.pixelsdb.pixels.common.physical.io.S3OutputStream;
 import io.pixelsdb.pixels.common.utils.ConfigFactory;
 import io.pixelsdb.pixels.common.utils.EtcdUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration;
-import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.http.apache.ApacheHttpClient;
 import software.amazon.awssdk.http.crt.AwsCrtAsyncHttpClient;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.*;
-import software.amazon.awssdk.services.s3.waiters.S3Waiter;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 
 import static io.pixelsdb.pixels.common.lock.EtcdAutoIncrement.GenerateId;
 import static io.pixelsdb.pixels.common.lock.EtcdAutoIncrement.InitId;
 import static io.pixelsdb.pixels.common.utils.Constants.S3_ID_KEY;
 import static io.pixelsdb.pixels.common.utils.Constants.S3_META_PREFIX;
-import static java.util.Objects.requireNonNull;
 
 /**
- * <p>
- * For S3, we assume that each table is stores as one bucket.
- * And all the pixels files in this table are store as separate objects
- * in the bucket. We do not use object key prefix for performance
- * considerations.
- * </p>
+ * For S3, we assume that each table is stored in a separate folder
+ * (i.e., a prefix or empty object in a bucket). And all the pixels
+ * files in this table are stored as individual objects in the folder.
  * <br/>
+ * @author hank
  * Created at: 20/08/2021
- * Author: hank
  */
-public class S3 implements Storage
+public final class S3 extends AbstractS3
 {
+    /*
+     * Most of the methods in this class are moved into AbstractS3.
+     */
+
     private static Logger logger = LogManager.getLogger(S3.class);
     private static String SchemePrefix = Scheme.s3.name() + "://";
 
-    private static int connectionTimeoutSec = 60;
-    private static int connectionAcquisitionTimeoutSec = 60;
-    private static int clientServiceThreads = 20;
-    private static int maxRequestConcurrency = 200;
-    private static int maxPendingRequests = 50_000;
-    private static S3Client s3;
-    private static S3AsyncClient s3Async;
-    private static S3AsyncClient s3Async1M;
-    private static S3AsyncClient s3Async10M;
-    private final static boolean enableCache;
     private final static boolean enableRequestDiversion;
+
+    private S3AsyncClient s3Async;
+    private S3AsyncClient s3Async1M;
+    private S3AsyncClient s3Async10M;
 
     static
     {
-        enableCache = Boolean.parseBoolean(ConfigFactory.Instance().getProperty("cache.enabled"));
-
         if (enableCache)
         {
             /**
@@ -96,21 +74,18 @@ public class S3 implements Storage
             InitId(S3_ID_KEY);
         }
 
-        connectionTimeoutSec = Integer.parseInt(
-                ConfigFactory.Instance().getProperty("s3.connection.timeout.sec"));
-        connectionAcquisitionTimeoutSec = Integer.parseInt(
-                ConfigFactory.Instance().getProperty("s3.connection.acquisition.timeout.sec"));
-        clientServiceThreads = Integer.parseInt(
-                ConfigFactory.Instance().getProperty("s3.client.service.threads"));
-        maxRequestConcurrency = Integer.parseInt(
-                ConfigFactory.Instance().getProperty("s3.max.request.concurrency"));
-        maxPendingRequests = Integer.parseInt(
-                ConfigFactory.Instance().getProperty("s3.max.pending.requests"));
-
         enableRequestDiversion = Boolean.parseBoolean(
                 ConfigFactory.Instance().getProperty("s3.enable.request.diversion"));
         logger.info("Request diversion enabled: " + enableRequestDiversion);
+    }
 
+    public static boolean isRequestDiversionEnabled()
+    {
+        return enableRequestDiversion;
+    }
+
+    public S3()
+    {
         String[] concurrencyAssign = null;
         if (enableRequestDiversion)
         {
@@ -118,7 +93,8 @@ public class S3 implements Storage
             logger.info("Request concurrency assignment: " + assign);
             concurrencyAssign = assign.split(":");
         }
-/*
+
+        /*
         s3Async = S3AsyncClient.builder()
                 .httpClientBuilder(NettyNioAsyncHttpClient.builder()
                         .connectionTimeout(Duration.ofSeconds(connectionTimeoutSec))
@@ -126,7 +102,8 @@ public class S3 implements Storage
                         .connectionAcquisitionTimeout(Duration.ofSeconds(connectionAcquisitionTimeoutSec))
                         .eventLoopGroup(SdkEventLoopGroup.builder().numberOfThreads(clientServiceThreads).build())
                         .maxConcurrency(maxRequestConcurrency).maxPendingConnectionAcquires(maxPendingRequests)).build();
-*/
+        */
+
         int maxConcurrency, maxConcurrency1M, maxConcurrency10M;
         if (enableRequestDiversion)
         {
@@ -179,14 +156,8 @@ public class S3 implements Storage
                 .maxConnections(maxRequestConcurrency)).build();
     }
 
-    public static boolean isRequestDiversionEnabled()
-    {
-        return enableRequestDiversion;
-    }
-
-    public S3() { }
-
-    private String getPathKey(String path)
+    @Override
+    protected String getPathKey(String path)
     {
         return S3_META_PREFIX + path;
     }
@@ -198,87 +169,6 @@ public class S3 implements Storage
             return key.substring(S3_META_PREFIX.length());
         }
         return null;
-    }
-
-    public static class Path
-    {
-        public String bucket = null;
-        public String key = null;
-        public boolean valid = false;
-        /**
-         * True if this path is folder.
-         * In S3, 'folder' is an empty object with its name ends with '/'.
-         * Besides that, we also consider the path that only contains a
-         * bucket name as a folder.
-         */
-        public boolean isFolder = false;
-
-        public Path(String path)
-        {
-            requireNonNull(path);
-            // remove the scheme header.
-            if (path.contains("://"))
-            {
-                path = path.substring(path.indexOf("://") + 3);
-            }
-            else if (path.startsWith("/"))
-            {
-                path = path.substring(1);
-            }
-            // the substring before the first '/' is the bucket name.
-            int slash = path.indexOf("/");
-            if (slash > 0)
-            {
-                this.bucket = path.substring(0, slash);
-                if (slash < path.length()-1)
-                {
-                    this.key = path.substring(slash + 1);
-                    this.isFolder = this.key.endsWith("/");
-                }
-                else
-                {
-                    // this is a bucket.
-                    this.isFolder = true;
-                }
-                this.valid = true;
-            }
-            else if (path.length() > 0)
-            {
-                // this is a bucket.
-                this.bucket = path;
-                this.isFolder = true;
-                this.valid = true;
-            }
-        }
-
-        public Path(String bucket, String key)
-        {
-            this.bucket = requireNonNull(bucket, "bucket is null");
-            this.key = key;
-            this.valid = true;
-            if (key != null)
-            {
-                this.isFolder = key.endsWith("/");
-            }
-            else
-            {
-                this.isFolder = true;
-            }
-        }
-
-        @Override
-        public String toString()
-        {
-            if (!this.valid)
-            {
-                return null;
-            }
-            if (this.key == null)
-            {
-                return this.bucket;
-            }
-            return this.bucket + "/" + this.key;
-        }
     }
 
     @Override
@@ -300,297 +190,6 @@ public class S3 implements Storage
                     "' already has a different scheme prefix than '" + SchemePrefix + "'.");
         }
         return SchemePrefix + path;
-    }
-
-    /**
-     * List status of a file or directory.
-     * Note that S3 does not support real directories, a directory / folder is
-     * an empty object with its name ends with '/'. So that of we want to list the
-     * status of the objects in a 'folder', the path must ends with '/', otherwise
-     * we can not filter out the status of the 'folder' itself from the returned result.
-     * @param path
-     * @return
-     * @throws IOException
-     */
-    @Override
-    public List<Status> listStatus(String path) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        ListObjectsV2Request.Builder builder = ListObjectsV2Request.builder().bucket(p.bucket);
-        if (p.key != null)
-        {
-            builder.prefix(p.key);
-        }
-        ListObjectsV2Request request = builder.build();
-        ListObjectsV2Response response = s3.listObjectsV2(request);
-        List<S3Object> objects = new ArrayList<>(response.keyCount());
-        while (response.isTruncated())
-        {
-            objects.addAll(response.contents());
-            request = builder.continuationToken(response.nextContinuationToken()).build();
-            response = s3.listObjectsV2(request);
-        }
-        objects.addAll(response.contents());
-        List<Status> statuses = new ArrayList<>();
-        Path op = new Path(path);
-        for (S3Object object : objects)
-        {
-            if (object.key().equals(p.key))
-            {
-                continue;
-            }
-            op.key = object.key();
-            statuses.add(new Status(op.toString(), object.size(), op.key.endsWith("/"), 1));
-        }
-        return statuses;
-    }
-
-    @Override
-    public List<String> listPaths(String path) throws IOException
-    {
-        return this.listStatus(path).stream().map(Status::getPath)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * For S3, the replication is always 1.
-     * @param path
-     * @return
-     * @throws IOException
-     */
-    @Override
-    public Status getStatus(String path) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        if (p.isFolder)
-        {
-            return new Status(p.toString(), 0, true, 1);
-        }
-
-        HeadObjectRequest request = HeadObjectRequest.builder().bucket(p.bucket).key(p.key).build();
-        try
-        {
-            HeadObjectResponse response = s3.headObject(request);
-            return new Status(p.toString(), response.contentLength(), false, 1);
-        } catch (Exception e)
-        {
-            throw new IOException("Failed to get object head of '" + path + "'", e);
-        }
-    }
-
-    @Override
-    public long getFileId(String path) throws IOException
-    {
-        requireNonNull(path, "path is null");
-        if (enableCache)
-        {
-            Path p = new Path(path);
-            if (!p.valid)
-            {
-                throw new IOException("Path '" + path + "' is not valid.");
-            }
-            // try to generate the id in etcd if it does not exist.
-            if (!this.existsOrGenIdSucc(p))
-            {
-                throw new IOException("Path '" + path + "' does not exist.");
-            }
-            KeyValue kv = EtcdUtil.Instance().getKeyValue(getPathKey(p.toString()));
-            return Long.parseLong(kv.getValue().toString(StandardCharsets.UTF_8));
-        }
-        else
-        {
-            // Issue #222: return an arbitrary id when cache is disable.
-            return path.hashCode();
-        }
-    }
-
-    @Override
-    public boolean mkdirs(String path) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        if (!p.isFolder)
-        {
-            throw new IOException("Path '" + path + "' is a directory, " +
-                    "the key for S3 directory (folder) must ends with '/'.");
-        }
-        if (this.existsInS3(p))
-        {
-            throw new IOException("Path '" + path + "' already exists.");
-        }
-
-        if (!this.existsInS3(new Path(p.bucket)))
-        {
-            CreateBucketRequest request = CreateBucketRequest.builder().bucket(p.bucket).build();
-            s3.createBucket(request);
-            S3Waiter waiter = s3.waiter();
-            HeadBucketRequest requestWait = HeadBucketRequest.builder()
-                    .bucket(p.bucket).build();
-            waiter.waitUntilBucketExists(requestWait);
-        }
-
-        if (p.key != null)
-        {
-            PutObjectRequest request = PutObjectRequest.builder().bucket(p.bucket).key(p.key).build();
-            s3.putObject(request, RequestBody.empty());
-
-            S3Waiter waiter = s3.waiter();
-            HeadObjectRequest requestWait = HeadObjectRequest.builder()
-                    .bucket(p.bucket).key(p.key).build();
-            waiter.waitUntilObjectExists(requestWait);
-        }
-        return true;
-    }
-
-    /**
-     * For S3, this open method is only used to read the data object
-     * fully and sequentially. And it will load the whole object into
-     * memory, so be careful for large objects.
-     * @param path
-     * @return
-     * @throws IOException
-     */
-    @Override
-    public DataInputStream open(String path) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        if (!this.existsInS3(p))
-        {
-            throw new IOException("Path '" + path + "' does not exist.");
-        }
-        return new DataInputStream(new S3InputStream(s3, p.bucket, p.key));
-    }
-
-    /**
-     * Open an output stream to write a file into S3.
-     * The access latency of cloud storage, such as S3, is high. If we are sure that the file path
-     * does not exist (e.g., the file name a UUID), we can skip file existence checking by setting
-     * overwrite to true.
-     * @param path
-     * @param overwrite
-     * @param bufferSize
-     * @return
-     * @throws IOException
-     */
-    @Override
-    public DataOutputStream create(String path, boolean overwrite, int bufferSize) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        if (!overwrite && this.existsInS3(p))
-        {
-            throw new IOException("Path '" + path + "' already exists.");
-        }
-        return new DataOutputStream(new S3OutputStream(s3, p.bucket, p.key, bufferSize));
-    }
-
-    @Override
-    public boolean delete(String path, boolean recursive) throws IOException
-    {
-        Path p = new Path(path);
-        if (!p.valid)
-        {
-            throw new IOException("Path '" + path + "' is not valid.");
-        }
-        if (!this.existsInS3(p))
-        {
-            // Issue #222: try to delete the file ids even if cache is disabled.
-            EtcdUtil.Instance().deleteByPrefix(getPathKey(p.toString()));
-            throw new IOException("Path '" + path + "' does not exist.");
-        }
-        if (p.isFolder)
-        {
-            List<Status> statuses = this.listStatus(path);
-            for (Status status : statuses)
-            {
-                // The ListObjects S3 API, which is used by listStatus, is already recursive.
-
-                Path sub = new Path(status.getPath());
-                DeleteObjectRequest request = DeleteObjectRequest.builder().bucket(sub.bucket).key(sub.key).build();
-                try
-                {
-                    s3.deleteObject(request);
-                } catch (Exception e)
-                {
-                    throw new IOException("Failed to delete object '" + sub.bucket + "/" + sub.key + "' from S3.", e);
-                }
-            }
-        }
-        else
-        {
-            DeleteObjectRequest request = DeleteObjectRequest.builder().bucket(p.bucket).key(p.key).build();
-            try
-            {
-                s3.deleteObject(request);
-            } catch (Exception e)
-            {
-                throw new IOException("Failed to delete object '" + p.bucket + "/" + p.key + "' from S3.", e);
-            }
-        }
-        // Issue #222: try to delete the file ids even if cache is disabled.
-        EtcdUtil.Instance().deleteByPrefix(getPathKey(p.toString()));
-        return true;
-    }
-
-    @Override
-    public boolean supportDirectCopy()
-    {
-        return true;
-    }
-
-    @Override
-    public boolean directCopy(String src, String dest) throws IOException
-    {
-        Path srcPath = new Path(src);
-        Path destPath = new Path(dest);
-        if (!srcPath.valid)
-        {
-            throw new IOException("Path '" + src + "' is not valid.");
-        }
-        if (!destPath.valid)
-        {
-            throw new IOException("Path '" + dest + "' is not valid.");
-        }
-        if (!this.existsInS3(srcPath))
-        {
-            throw new IOException("Path '" + src + "' does not exist.");
-        }
-        if (this.existsInS3(destPath))
-        {
-            throw new IOException("Path '" + dest + "' already exists.");
-        }
-        CopyObjectRequest copyReq = CopyObjectRequest.builder()
-                .copySource(srcPath.toString())
-                .destinationBucket(destPath.bucket)
-                .destinationKey(destPath.key)
-                .build();
-        try
-        {
-            s3.copyObject(copyReq);
-            return true;
-        }
-        catch (RuntimeException e)
-        {
-            throw new IOException("Failed to copy object from '" + src + "' to '" + dest + "'", e);
-        }
     }
 
     @Override
@@ -615,50 +214,7 @@ public class S3 implements Storage
     }
 
     @Override
-    public boolean exists(String path) throws IOException
-    {
-        return this.existsInS3(new Path(path));
-    }
-
-    /**
-     * If a file or directory exists in S3.
-     * @param path
-     * @return
-     * @throws IOException
-     */
-    private boolean existsInS3(Path path) throws IOException
-    {
-        if (!path.valid)
-        {
-            throw new IOException("Path '" + path.toString() + "' is not valid.");
-        }
-
-        try
-        {
-            if (path.key == null)
-            {
-                HeadBucketRequest request = HeadBucketRequest.builder().bucket(path.bucket).build();
-                s3.headBucket(request);
-            }
-            else
-            {
-                HeadObjectRequest request = HeadObjectRequest.builder().bucket(path.bucket).key(path.key).build();
-                s3.headObject(request);
-            }
-            return true;
-        } catch (Exception e)
-        {
-            if (e instanceof NoSuchKeyException ||
-            e instanceof NoSuchBucketException)
-            {
-
-                return false;
-            }
-            throw new IOException("Failed to check the existence of '" + path + "'", e);
-        }
-    }
-
-    private boolean existsOrGenIdSucc(Path path) throws IOException
+    protected boolean existsOrGenIdSucc(Path path) throws IOException
     {
         if (!enableCache)
         {
@@ -680,23 +236,6 @@ public class S3 implements Storage
             return true;
         }
         return false;
-    }
-
-    @Override
-    public boolean isFile(String path) throws IOException
-    {
-        return !(new Path(path).isFolder);
-    }
-
-    @Override
-    public boolean isDirectory(String path) throws IOException
-    {
-        return new Path(path).isFolder;
-    }
-
-    public S3Client getClient()
-    {
-        return s3;
     }
 
     public S3AsyncClient getAsyncClient()
