+++ pixels-core/src/test/java/io/pixelsdb/pixels/core/predicate/TestPredicate.java
@@ -30,6 +30,7 @@ import io.pixelsdb.pixels.core.vector.VectorizedRowBatch;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.SortedMap;
 import java.util.TreeMap;
 
@@ -141,5 +142,22 @@ public class TestPredicate
         recordReader.close();
         pixelsReader.close();
         pixelsWriter.close();
+    }
+
+    @Test
+    public void testStringDecode()
+    {
+        String str = "hello world";
+        byte[] bytes = str.getBytes(StandardCharsets.UTF_8);
+        assert bytes.length == 11;
+        bytes = str.getBytes(StandardCharsets.ISO_8859_1);
+        assert bytes.length == 11;
+        str = "你好世界";
+        bytes = str.getBytes(StandardCharsets.UTF_8);
+        System.out.println(bytes.length);
+        for (byte b : bytes)
+        {
+            System.out.println(b);
+        }
     }
 }
